Update
Always visible shader
+
Changed some objects to static
+
The player can be seen behind obstacles
+
Improving the machine increases the area for collecting boxes.

+++ Prototype_Casual/Assets/Scripts/ShopData.cs
@@ -28,7 +28,7 @@ namespace ShopSystem
     {
         public int unlockCost;
         public int speed;
-        public int acceleration;
+        public float acceleration;
 
     }
 }
